Fix a compile issue converting 'long long' to int64_t (couldn't reproduce, but Isaac said this will work)
git-svn-id: http://svn.mythtv.org/svn/trunk@7957 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/programs/mythtranscode/mpeg2fix.h
@@ -193,7 +193,7 @@ class MPEG2fixup
 
     MPEG2ptsDeltaList *ptsDelta;
 
-    QMap<int64_t, int> delMap;
+    QMap<long long, int> delMap;
 
     pthread_t thread;
